Change Cuda pinned memory allocation to BFC allocator. Move the GPU-neutral code to common_runtime. Change: 117591254
