+++ tensorflow/core/common_runtime/gpu/gpu_debug_allocator.h
@@ -20,7 +20,7 @@ limitations under the License.
 #include <string>
 #include <unordered_map>
 
-#include "tensorflow/core/common_runtime/gpu/visitable_allocator.h"
+#include "tensorflow/core/common_runtime/visitable_allocator.h"
 #include "tensorflow/core/platform/macros.h"
 #include "tensorflow/core/platform/stream_executor.h"
 #include "tensorflow/core/platform/types.h"
